api/requirements.txt
fastapi==0.115.2
keybert
pydantic
pymongo
python-dotenv
transformers
uvicorn
starlette
